Plot bycatch estimation CV vs. observer coverage

plot_cv plots projected bycatch estimation CVs vs observer coverage, and returns minimum observer coverage needed to achieve user-specified target CV and percentile.

plot_cv(te, bpue, d = 2, targetcv = 0.3, showplot = TRUE,
  silent = FALSE, ...)

Arguments

te

an integer greater than 1. Total effort in fishery (e.g., trips or sets).

bpue

a positive number. Bycatch per unit effort.

d

a number greater than or equal to 1. Dispersion index. The dispersion index corresponds to the variance-to-mean ratio of effort-unit-level bycatch, so d = 1 corresponds to Poisson-distributed bycatch, and d > 1 to overdispersed bycatch.

targetcv

a non-negative number less than 1. Target CV (as a proportion). If set to 0, no corresponding minimum observer coverage will be highlighted or returned.

showplot

logical. If FALSE, plotting is suppressed.

silent

logical. If TRUE, print output to terminal is suppressed.

...

additional arguments for compatibility with Shiny.

Value

If targetcv is non-zero, a list with one component:

targetoc

minimum observer coverage in terms of percentage.

Returned invisibly.

Details

Caveat: plot_cv assumes that (1) observer coverage is representative, (2) bycatch (bpue) is in terms of individuals (not weight) per unit effort, and (3) the specified dispersion index reflects the highest level of any hierarchical variance (e.g., using dispersion index at trip level if greater than that at set level). Violating these assumptions will likely result in negatively biased projections of the observer coverage needed to meet a specified objective. More conservative (higher) projections can be obtained by using a higher dispersion index d. Users may want to explore uncertainty in dispersion index and in bycatch per unit effort by varying those inputs.