Simulate CV response to observer coverage

sim_cv_obscov simulates bycatch estimation CVs for a range of observer coverage levels, given bycatch rate, dispersion index, and total fishery effort. The resulting object is required by plot_cv_obscov for visualization.

sim_cv_obscov(te, bpue, d = 2, nsim = 1000, ...)

Arguments

te

an integer greater than one. Total independent units of effort in fishery (e.g., trips or sets, not landings).

bpue

a positive number. Bycatch per unit effort.

d

a number greater than or equal to 1. Dispersion index. The dispersion index corresponds to the variance-to-mean ratio of effort-unit-level bycatch, so \(d = 1\) corresponds to Poisson- distributed bycatch, and \(d > 1\) corresponds to overdispersed bycatch.

nsim

a positive integer. Number of simulations to run.

...

additional arguments for compatibility with Shiny.

Value

A list with components:

simsum

a tibble with one row per observer coverage level and the following fields: simulated proportion observer coverage (pobs), number of observed trips/sets (nobs), and bycatch estimation CV (cvsim).

simdat

a tibble with one row per simulation and the following fields: simulated proportion observer coverage (pobs), number of observed trips/sets (nobs), true (realized) bycatch per unit effort (tbpue), observed bycatch per unit effort (obpue), and error of observed bycatch per unit effort (oberr = obpue - tbpue).

bpue

the nominal bycatch per unit effort used in the simulations.

d

the dispersion index used in the simulations.

Details

sim_cv_obscov runs nsim simulations per level of observer coverage, from the larger of 0.1% or one trip/set to 100%. Simulated bycatch estimates are calculated as mean observed bycatch per unit effort. CV at each observer coverage level is calculated as the square root of mean square estimation error divided by "true" bycatch (bpue).

Warning: Large total effort (>100K trips/sets) may require several minutes of execution time. Increasing nsim from the default of 1000 will also increase execution time.

Caveat: sim_cv_obscov assumes representative observer coverage and independent units of effort, i.e., no hierarchical sources of variance such as vessel- or trip-level variation if sets or hauls are used. Violating these assumptions will likely result in negatively biased projections of bycatch estimation CV for a given level of observer coverage. More conservative projections can be obtained by using higher-level units of effort (e.g., bpue as mean bycatch per trip instead of bycatch per trip/set, and te as number of trips instead of number of trips/sets). Landings (number or weight) do not represent independent sampling units unless an average independent fishing effort unit such as a trip lands less than or equal to one measurement unit of landings (e.g., <= one metric ton per trip).