Plot probability of positive bycatch vs observer coverage

plot_probposobs plots (1) probability of observing at least one bycatch event vs observer coverage and (2) probability of any bycatch occurring in total effort, given total effort in trips/sets, bycatch per unit effort, and dispersion index.

plot_probposobs(
  te,
  bpue,
  d = 2,
  targetppos = 95,
  showplot = TRUE,
  silent = FALSE
)

Arguments

te
an integer greater than 1. Total effort in fishery (trips/sets).

bpue
a positive number. Bycatch per unit effort.

d
a number greater than or equal to 1. Dispersion index. The dispersion index corresponds to the variance-to-mean ratio of effort-unit-level bycatch, so \(d = 1\) corresponds to Poisson- distributed bycatch, and \(d > 1\) corresponds to overdispersed bycatch.

targetppos
a non-negative number less than or equal to 100. Target probability of positive observed bycatch (as percentage), given positive bycatch in total effort. If 0, no corresponding minimum observer coverage will be highlighted.

showplot
logical. If plot = FALSE, plotting is suppressed.

silent
logical. If silent = TRUE, print output to terminal is suppressed.

Value

A list with components:

pobs

minimum observer coverage in terms of percentage.

nobs

corresponding observed effort.

ppos.te

probability of any bycatch occurring in total effort

Returned invisibly.
Details

Probabilities are based on the probability density function for the corresponding Poisson or negative binomial distribution.

The probability that any bycatch occurs in the given total effort is shown by the horizontal black dotted line. The conditional probability of observing any bycatch if it occurs is shown by the solid black line. The product of these first two probabilities gives the absolute probability of observing any bycatch (dashed black line).The minimum observer coverage to achieve the target obability of observing bycatch if it occurs (x-axis value of red star) is where the conditional bycatch detection probability (solid black line) intersects with the target probability (red dash-dot line).

Caveat: plot_probposobs assumes representative observer coverage and no hierarchical sources of variance (e.g., vessel- or trip-level variation). Violating these assumptions will likely result in positively biased projections of the probability of observing bycatch at a given level of observer coverage. More conservative projections can be obtained by using higher-level units of effort (e.g., bpue as mean bycatch per trip instead of bycatch per trip/set, and te as number of trips instead of number of trips/sets).