Plot probability of positive bycatch vs observer coverage

plot_probposobs plots (1) probability of observing at least one bycatch event vs observer coverage and (2) probability of any bycatch occurring in total fishery effort, given total fishery effort, bycatch per unit effort, and dispersion index. The function returns returns minimum observer coverage needed to achieve user-specified probability of observing bycatch if it occurs.

plot_probposobs(te, bpue, d = 2, targetppos = 95, showplot = TRUE,
  silent = FALSE, ...)

Arguments

te
an integer greater than 1. Total effort in fishery (e.g., trips or sets).

bpue
a positive number. Bycatch per unit effort.

d
a number greater than or equal to 1. Dispersion index. The dispersion index corresponds to the variance-to-mean ratio of effort-unit-level bycatch, so d = 1 corresponds to Poisson-distributed bycatch, and d > 1 to overdispersed bycatch.

targetppos
a non-negative number less than or equal to 100. Target probability of positive observed bycatch (as percentage), given positive bycatch in total effort. If set to 0, no corresponding minimum observer coverage will be highlighted or returned.

showplot
logical. If FALSE, plotting is suppressed.

silent
logical. If TRUE, print output to terminal is suppressed.

...
additional arguments for compatibility with Shiny.

Value

A list with two components:

targetoc

minimum observer coverage in terms of percentage.

ppos.te

probability of any bycatch occurring in total effort

Returned invisibly.
Details

Probabilities are based on the probability density function for the corresponding Poisson or negative binomial distribution.

The conditional probability of observing any bycatch if it occurs (solid black line) is obtained by dividing the absolute probability of observing any bycatch (black dashed line) by the probability that any bycatch occurs in the given total effort (horizontal black dotted line). The minimum observer coverage to achieve the target probability of observing bycatch if it occurs (x-axis value of red star) is where the conditional bycatch detection probability (solid black line) intersects with the target probability (red dash-dot line).

Caveat: plot_probposobs assumes that (1) observer coverage is representative, (2) bycatch (bpue) is in terms of individuals (not weight) per unit effort, and (3) the specified dispersion index reflects the highest level of any hierarchical variance (e.g., using dispersion index at trip level if greater than that at set level). Violating these assumptions will likely result in positively biased projections of the probability of observing bycatch at a given level of observer coverage. More conservative projections can be obtained by using a higher dispersion index d. Users may want to explore uncertainty in dispersion index and in bycatch per unit effort by varying those inputs.