Plot upper confidence limit of total bycatch given none observed

plot_uclnegobs plots upper confidence limit of total bycatch vs observer coverage when no bycatch is observed, given total fishery effort, dispersion index, and confidence level.

plot_uclnegobs(
  te,
  d = 2,
  cl = 95,
  targetucl = 0,
  fixedoc = 0,
  ymax = 100,
  showplot = TRUE,
  silent = FALSE
)

Arguments

te

an integer greater than 1. Total effort in fishery (e.g., trips or sets).

d

a number greater than or equal to 1. Dispersion index. The dispersion index corresponds to the variance-to-mean ratio of effort-unit-level bycatch, so \(d = 1\) corresponds to Poisson- distributed bycatch, and \(d > 1\) corresponds to overdispersed bycatch.

cl

a non-negative number less than or equal to 100. Confidence level for upper confidence limit of total bycatch (as percentage), given no bycatch observed.

targetucl

a non-negative number. Target maximum upper confidence limit for total bycatch given zero bycatch observed. If 0, no corresponding minimum observer coverage will be highlighted.

fixedoc

a non-negative number between 0 and 100. Percent observer coverage for which to return ucl value.

ymax

a positive number. Upper limit for y-axis of plot.

showplot

logical. If plot = FALSE, plotting is suppressed.

silent

logical. If silent = TRUE, print output to terminal is suppressed.

Value

A list with components:

ucldat

a tibble with the following fields for each coverage level included: number of observed trips or sets (nobs), proportion observer coverage (pobs), upper confidence limit of total bycatch given none observed (ucl), and finite population correction (fpc) used in calculating ucl.

targetucl

specified target maximum upper confidence limit of bycatch.

targetoc

minimum observer coverage (as percentage) for which upper confidence limit of bycatch is targetucl when none observed.

targetnoc

minimum observer coverage (as effort) for which upper confidence limit of bycatch is targetucl when none observed.

fixedoc

specified percentage observer coverage for which upper confidence limit of bycatch is returned.

fixednoc

observer coverage (as effort) corresponding to fixedoc.

fixedoc.ucl

upper confidence limit of total bycatch corresponding to zero bycatch observed in fixedoc coverage.

te

specified total effort.

d

specified dispersion index.

cl

specified confidence level.

Returned invisibly.

Details

Upper confidence limits are based on the probability density function for the corresponding Poisson or negative binomial distribution. Upper confidence limits based on d+/-1 (as allowed by specification of d) are also plotted. If fixedoc specified, corresponding upper confidence limit is provided in printed output and returned object, but not in plot.

Caveat: plot_uclnegobs assumes that (1) observer coverage is representative, (2) bycatch is in terms of individuals (not weight) per unit effort,and (3) the specified dispersion index reflects the highest level of any hierarchical variance (e.g., using dispersion index at trip level if greater than that at set level). Violating these assumptions will likely result in negatively biased projections of the upper confidence limit of total bycatch given zero observed. More conservative projections can be obtained by using a higher dispersion index d. Users may want to explore uncertainty in dispersion index and in bycatch per unit effort by varying those inputs.